Angular 8 Refactor

+++ src/app/components/settings/menu-settings/menu-settings.component.ts
@@ -6,6 +6,7 @@ import { Report } from '../../../mocks/report';
 import { MessageService } from '../../../providers/message.service';
 import { LogService, logType } from '../../../services/log.service';
 import { MainService } from '../../../services/main.service';
+import { EntityStoreService } from '../../../services/entity-store.service';
 
 @Component({
   selector: 'app-menu-settings',
@@ -14,6 +15,7 @@ import { MainService } from '../../../services/main.service';
 })
 
 export class MenuSettingsComponent implements OnInit {
+
   categories: Array<Category>;
   sub_categories: Array<SubCategory>;
   products: Array<Product>;
@@ -35,15 +37,17 @@ export class MenuSettingsComponent implements OnInit {
   subCats: Array<SubCategory> = [];
   printers: Array<Printer>;
   productSpecs: Array<ProductSpecs>;
+  categoryNames: Map<string, string> = new Map();
 
   @ViewChild('catDetails') catDetails: NgForm;
   @ViewChild('subCatForm') subCatForm: NgForm;
   @ViewChild('productForm') productForm: NgForm;
   @ViewChild('categoryForm') categoryForm: NgForm;
   @ViewChild('recipesForm') recipesForm: NgForm;
-  @ViewChild('productTypeSelect') productTypeSelect: ElementRef
+  @ViewChild('productTypeSelect') productTypeSelect: ElementRef;
 
-  constructor(private mainService: MainService, private messageService: MessageService, private logService: LogService) {
+  constructor(private mainService: MainService, private messageService: MessageService,
+              private logService: LogService, private entityStoreService: EntityStoreService) {
     this.fillData();
   }
 
@@ -90,15 +94,8 @@ export class MenuSettingsComponent implements OnInit {
     }
   }
 
-  getProductsBySubCat(id) {
-    this.mainService.getAllBy('products', { subcat_id: id }).then(result => {
-      this.products = result.docs;
-      this.products = this.products.sort((a, b) => a.price - b.price);
-    });
-  }
-
   addCategory(categoryForm: NgForm) {
-    let form = categoryForm.value;
+    const form = categoryForm.value;
     if (!form.name) {
       this.messageService.sendMessage('Kategori Adı Belirtmelisiniz');
       return false;
@@ -106,7 +103,7 @@ export class MenuSettingsComponent implements OnInit {
     if (!form.printer) {
       form.printer = '';
     }
-    let schema = new Category(form.name, form.description, 1, form.printer, 0, form.tags);
+    const schema = new Category(form.name, form.description, 1, form.printer, 0, form.tags);
     this.mainService.addData('categories', schema).then(() => {
       this.fillData();
       this.messageService.sendMessage('Kategori Oluşturuldu');
@@ -116,7 +113,7 @@ export class MenuSettingsComponent implements OnInit {
   }
 
   updateCategory(catDetails) {
-    let form = catDetails.value;
+    const form = catDetails.value;
     this.mainService.updateData('categories', form._id, form).then(() => {
       this.fillData();
       this.messageService.sendMessage('Kategori Düzenlendi');
@@ -125,44 +122,50 @@ export class MenuSettingsComponent implements OnInit {
   }
 
   removeCategory(id) {
-    let isOk = confirm('Kategoriyi Silmek Üzeresiniz. Kategoriye Dahil Olan Ürünlerde Silinecektir.');
+    const isOk = confirm('Kategoriyi Silmek Üzeresiniz. Kategoriye Dahil Olan Ürünlerde Silinecektir.');
     if (isOk) {
       this.mainService.getAllBy('products', { cat_id: id }).then(result => {
-        let data = result.docs
+        const data = result.docs;
         if (data.length > 0) {
-          for (let prop in data) {
-            this.mainService.removeData('products', data[prop]._id);
-            this.mainService.getAllBy('reports', { connection_id: data[prop]._id }).then(res => {
-              if (res.docs.length > 0)
-                this.mainService.removeData('reports', res.docs[0]._id);
-            });
-            this.mainService.getAllBy('recipes', { product_id: data[prop]._id }).then(res => {
-              if (res.docs.length > 0)
-                this.mainService.removeData('recipes', res.docs[0]._id);
-            });
+          for (const prop in data) {
+            if (Object.prototype.hasOwnProperty.call(data, prop)) {
+              this.mainService.removeData('products', data[prop]._id);
+              this.mainService.getAllBy('reports', { connection_id: data[prop]._id }).then(res => {
+                if (res.docs.length > 0) {
+                  this.mainService.removeData('reports', res.docs[0]._id);
+                }
+              });
+              this.mainService.getAllBy('recipes', { product_id: data[prop]._id }).then(res => {
+                if (res.docs.length > 0) {
+                  this.mainService.removeData('recipes', res.docs[0]._id);
+                }
+              });
+            }
           }
         }
-      });
-      this.mainService.getAllBy('sub_categories', { cat_id: id }).then(res => {
-        let data = res.docs;
-        if (data.length > 0) {
-          for (let prop in data) {
-            this.mainService.removeData('sub_categories', data[prop]._id);
+        this.mainService.getAllBy('sub_categories', { cat_id: id }).then(res => {
+          const subCatData = res.docs;
+          if (subCatData.length > 0) {
+            for (const prop in subCatData) {
+              if (Object.prototype.hasOwnProperty.call(subCatData, prop)) {
+                this.mainService.removeData('sub_categories', subCatData[prop]._id);
+              }
+            }
           }
-        }
-      });
-      this.mainService.removeData('categories', id).then(() => {
-        this.selectedCat = undefined;
-        this.messageService.sendMessage('Kategori ve Ürünler Silindi');
-        this.fillData();
+        });
+        this.mainService.removeData('categories', id).then(() => {
+          this.selectedCat = undefined;
+          this.messageService.sendMessage('Kategori ve Ürünler Silindi');
+          this.fillData();
+        });
       });
     }
   }
 
   addSubCategory(subCatForm: NgForm) {
-    let form = subCatForm.value;
-    if (form._id == '' || form._id == undefined || form._id == null) {
-      let schema = new SubCategory(this.selectedCat._id, form.name, form.description, 1);
+    const form = subCatForm.value;
+    if (form._id === '' || form._id === undefined || form._id === null) {
+      const schema = new SubCategory(this.selectedCat._id, form.name, form.description, 1);
       this.mainService.addData('sub_categories', schema).then(res => {
         this.getCategory(this.selectedCat);
       });
@@ -194,28 +197,34 @@ export class MenuSettingsComponent implements OnInit {
   }
 
   addProduct(productForm: NgForm) {
-    let form = productForm.value;
+    const form = productForm.value;
     if (!form.name || !form.cat_id || !form.price || !form.tax_value) {
       this.messageService.sendMessage('Gerekli Alanları Doldurmalısınız');
       return false;
     }
-    if (form.type == 2 && this.oldRecipes.length == 0) {
-      if (form.type == 2 && this.productRecipe.length == 0) {
+    if (form.type === 2 && this.oldRecipes.length === 0) {
+      if (this.productRecipe.length === 0) {
         this.messageService.sendMessage('Stok Girişi Yapmalısınız!');
         return false;
       }
     }
-    let schema = new Product(form.cat_id, form.type, form.description, form.name, form.price, 1, form.tax_value, form.barcode, form.notes, form.subcat_id, this.productSpecs, form._id, form._rev);
-    if (form._id == undefined) {
+    const schema = new Product(form.cat_id, form.type, form.description, form.name, form.price, 1,
+      form.tax_value, form.barcode, form.notes, form.subcat_id,
+      this.productSpecs, form._id, form._rev);
+    if (form._id === undefined) {
       this.mainService.addData('products', schema).then((response) => {
-        this.mainService.addData('reports', new Report('Product', response.id, 0, 0, 0, [0, 0, 0, 0, 0, 0, 0], [0, 0, 0, 0, 0, 0, 0], [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], new Date().getMonth(), new Date().getFullYear(), schema.name, Date.now())).then(res => {
-          this.logService.createLog(logType.PRODUCT_CREATED, res.id, `${form.name} adlı Ürün Oluşturuldu`)
+        const weeklyData = [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+          0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
+        const weekly = [0, 0, 0, 0, 0];
+        const report = new Report('Product', response.id, 0, 0, 0, weekly, weekly, weeklyData, weeklyData,
+          new Date().getMonth(), new Date().getFullYear(), schema.name, Date.now());
+        this.mainService.addData('reports', report).then(res => {
+          this.logService.createLog(logType.PRODUCT_CREATED, res.id, `${form.name} adlı Ürün Oluşturuldu`);
         });
         if (this.productRecipe.length > 0) {
-          let schema = new Recipe(response.id, this.productRecipe);
-          this.mainService.addData('recipes', schema);
+          const recipeSchema = new Recipe(response.id, this.productRecipe);
+          this.mainService.addData('recipes', recipeSchema);
         }
-        // this.fillData();
         this.messageService.sendMessage('Ürün Oluşturuldu');
       });
     } else {
@@ -223,9 +232,9 @@ export class MenuSettingsComponent implements OnInit {
         if (res.ok) {
           this.logService.createLog(logType.PRODUCT_UPDATED, res.id, `${form.name} adlı Ürün Güncellendi`);
           if (this.productRecipe.length > 0) {
-            if (this.recipe.length == 0) {
-              let schema = new Recipe(form._id, this.productRecipe);
-              this.mainService.addData('recipes', schema);
+            if (this.recipe.length === 0) {
+              const recipeSchema = new Recipe(form._id, this.productRecipe);
+              this.mainService.addData('recipes', recipeSchema);
             } else {
               this.productRecipe = this.productRecipe.concat(this.recipe);
               this.mainService.updateData('recipes', this.recipeId, { recipe: this.productRecipe });
@@ -254,15 +263,15 @@ export class MenuSettingsComponent implements OnInit {
       this.mainService.getAllBy('sub_categories', { cat_id: result.cat_id }).then(res => {
         this.subCats = res.docs;
       });
-      result.note = "";
-      if (result.specifies == undefined || result.specifies == '') {
+      result.note = '';
+      if (result.specifies === undefined || result.specifies === '') {
         result.specifies = [];
       }
       this.productSpecs = result.specifies;
-      if (result.subcat_id == undefined) {
-        result.subcat_id = "";
+      if (result.subcat_id === undefined) {
+        result.subcat_id = '';
       }
-      if(!result.notes){
+      if (!result.notes) {
         result.notes = '';
       }
       this.productType = result.type;
@@ -273,12 +282,19 @@ export class MenuSettingsComponent implements OnInit {
       if (result.docs.length > 0) {
         this.hasRecipe = true;
         this.recipeId = result.docs[0]._id;
-        let recipes = result.docs[0].recipe;
+        const recipes = result.docs[0].recipe;
         this.recipe = result.docs[0].recipe;
-        for (let prop in recipes) {
-          this.mainService.getData('stocks', recipes[prop].stock_id).then(result => {
-            this.oldRecipes.push({ id: recipes[prop].stock_id, name: result.name, amount: recipes[prop].amount, unit: result.unit });
-          });
+        for (const prop in recipes) {
+          if (Object.prototype.hasOwnProperty.call(recipes, prop)) {
+            this.mainService.getData('stocks', recipes[prop].stock_id).then(stockResult => {
+              this.oldRecipes.push({
+                id: recipes[prop].stock_id,
+                name: stockResult.name,
+                amount: recipes[prop].amount,
+                unit: stockResult.unit
+              });
+            });
+          }
         }
       } else {
         this.hasRecipe = false;
@@ -286,53 +302,34 @@ export class MenuSettingsComponent implements OnInit {
     });
   }
 
-  removeProduct() {
-    let isOk = confirm('Ürünü Silmek Üzerisiniz..');
-    if (isOk) {
-      this.mainService.removeData('products', this.selectedId).then((result) => {
-        this.logService.createLog(logType.PRODUCT_DELETED, result.id, `${this.productForm.value.name} adlı Ürün Silindi`);
-        this.mainService.getAllBy('reports', { connection_id: result.id }).then(res => {
-          if (res.docs.length > 0)
-            this.mainService.removeData('reports', res.docs[0]._id);
-        });
-        this.mainService.getAllBy('recipes', { product_id: result.id }).then(res => {
-          if (res.docs.length > 0)
-            this.mainService.removeData('recipes', res.docs[0]._id);
-        });
-        this.messageService.sendMessage('Ürün Silindi!');
-        // this.fillData();
-      });
-    }
-  }
-
   setProductType(value) {
     this.productType = value;
-    if (value == 2) {
+    if (value === 2) {
       if (this.oldRecipes.length > 1) {
-        let isOK = confirm('Manuel Stok tipi için tek bir Stok kaydı girebilirsiniz. 2. Stok ve Ürün Durumları silinecektir.');
+        const isOK = confirm('Manuel Stok tipi için tek bir Stok kaydı girebilirsiniz. 2. Stok ve Ürün Durumlarını silinecektir.');
         if (isOK) {
-          let recipeWillLost = this.oldRecipes.pop();
+          const recipeWillLost = this.oldRecipes.pop();
           this.removeRecipe('old', recipeWillLost.id);
         } else {
           this.productType = 1;
           this.productTypeSelect.nativeElement.value = 1;
         }
       }
-      if (this.productRecipe.length > 0) {
-        this.productRecipe.pop();
-        this.recipesTable.pop();
-      }
+    }
+    if (this.productRecipe.length > 0) {
+      this.productRecipe.pop();
+      this.recipesTable.pop();
       this.productSpecs = [];
     }
   }
 
   addSpecies(speciesForm) {
-    let form = speciesForm.value;
-    if (form.spec_name == '' || form.spec_name == null || form.spec_price <= 0 || form.spec_price == undefined) {
+    const form = speciesForm.value;
+    if (form.spec_name === '' || form.spec_name === null || form.spec_price <= 0 || form.spec_price === undefined) {
       this.messageService.sendMessage('Durum Notu ve Fiyatı Boş Bırakılamaz');
       return false;
     }
-    let spec = new ProductSpecs(form.spec_name, form.spec_price);
+    const spec = new ProductSpecs(form.spec_name, form.spec_price);
     this.productSpecs.push(spec);
     speciesForm.reset();
   }
@@ -342,53 +339,52 @@ export class MenuSettingsComponent implements OnInit {
   }
 
   addRecipe(recipesForm) {
-    let form = recipesForm.value;
+    const form = recipesForm.value;
     if (!form.stock_id) {
       this.messageService.sendMessage('Stok Seçimi Yapmalısınız!');
       return false;
     }
-    if (!form.amount || form.amount == 0) {
+    if (!form.amount || form.amount === 0) {
       this.messageService.sendMessage('Birim Miktarı Girmelisiniz!');
       return false;
     }
-    let item = new Ingredient(form.stock_id, form.amount);
+    const item = new Ingredient(form.stock_id, form.amount);
     this.mainService.getData('stocks', form.stock_id).then(result => {
       // if (form.amount < result.left_total) {
-      if (this.productRecipe.find(item => item.stock_id == form.stock_id)) {
+      if (this.productRecipe.find(ingredient => ingredient.stock_id === form.stock_id)) {
         this.messageService.sendMessage('İçerik daha önce tanımlanmış!');
-      } else if (this.oldRecipes.find(item => item.id == form.stock_id)) {
+      } else if (this.oldRecipes.find(oldRecipe => oldRecipe.id === form.stock_id)) {
         this.messageService.sendMessage('İçerik daha önce tanımlanmış!');
       } else {
         this.productRecipe.push(item);
         this.recipesTable.push({ id: form.stock_id, name: this.stockName, amount: form.amount, unit: this.stockUnit });
         this.hasRecipe = true;
       }
-      // } else {
-      //   this.messageService.sendMessage('Elinizdeki  kalan stok miktarından fazla giremezsiniz.')
-      // }
-      recipesForm.reset();
     });
+    recipesForm.reset();
   }
 
   removeRecipe(type, id) {
     switch (type) {
       case 'new':
-        this.productRecipe = this.productRecipe.filter(item => item.stock_id != id);
-        this.recipesTable = this.recipesTable.filter(item => item.id != id);
-        if (this.recipesTable.length == 0) {
-          if (this.oldRecipes.length == 0) {
+        this.productRecipe = this.productRecipe.filter(item => item.stock_id !== id);
+        this.recipesTable = this.recipesTable.filter(item => item.id !== id);
+        if (this.recipesTable.length === 0) {
+          if (this.oldRecipes.length === 0) {
             this.hasRecipe = false;
           }
         }
         break;
       case 'old':
-        this.recipe = this.recipe.filter(item => item.stock_id != id);
-        this.productRecipe = this.productRecipe.filter(item => item.stock_id != id);
-        this.oldRecipes = this.oldRecipes.filter(item => item.id != id);
-        if (this.oldRecipes.length == 0) {
-          if (this.recipesTable.length == 0) {
+        this.recipe = this.recipe.filter(item => item.stock_id !== id);
+        this.productRecipe = this.productRecipe.filter(item => item.stock_id !== id);
+        this.oldRecipes = this.oldRecipes.filter(item => item.id !== id);
+        if (this.oldRecipes.length === 0) {
+          if (this.recipesTable.length === 0) {
             this.hasRecipe = false;
           }
+        }
+        if (this.recipe.length === 0) {
           this.mainService.removeData('recipes', this.recipeId);
         } else {
           this.mainService.updateData('recipes', this.recipeId, { recipe: this.recipe });
@@ -407,7 +403,7 @@ export class MenuSettingsComponent implements OnInit {
   }
 
   filterProducts(value: string) {
-    let regexp = new RegExp(value, 'i');
+    const regexp = new RegExp(value, 'i');
     this.mainService.getAllBy('products', { name: { $regex: regexp } }).then(res => {
       this.products = res.docs;
       this.products = this.products.sort((a, b) => a.price - b.price);
@@ -416,7 +412,7 @@ export class MenuSettingsComponent implements OnInit {
 
   setDefault() {
     this.productRecipe = [];
-    this.productSpecs = []
+    this.productSpecs = [];
     this.recipesTable = [];
     this.oldRecipes = [];
     this.recipe = [];
@@ -435,12 +431,24 @@ export class MenuSettingsComponent implements OnInit {
   }
 
   fillData() {
-    this.mainService.getAllBy('categories', {}).then(result => this.categories = result.docs);
+    this.mainService.getAllBy('categories', {}).then(result => {
+      this.categories = result.docs;
+    });
     this.mainService.getAllBy('products', {}).then(result => {
       this.products = result.docs;
       this.products = this.products.sort((a, b) => a.price - b.price);
+
+      // Resolve category names
+      const categoryIds = this.products.map(p => p.cat_id).filter(id => id);
+      this.entityStoreService.resolveEntities('categories', categoryIds).then(resolved => {
+        this.categoryNames = resolved;
+      });
     });
-    this.mainService.getAllBy('stocks', {}).then(result => this.stocks = result.docs);
-    this.mainService.getAllBy('settings', { key: 'Printers' }).then(res => this.printers = res.docs[0].value);
+    this.mainService.getAllBy('stocks', {}).then(result => {
+      this.stocks = result.docs;
+    });
+    this.mainService.getAllBy('settings', { key: 'Printers' }).then(res => {
+      this.printers = res.docs[0].value;
+    });
   }
 }
