+++ src/app/components/settings/restaurant-settings/restaurant-settings.component.ts
@@ -5,6 +5,7 @@ import { Floor, FloorSpecs, Table } from '../../../mocks/table';
 import { MessageService } from '../../../providers/message.service';
 import { LogService, logType } from '../../../services/log.service';
 import { MainService } from '../../../services/main.service';
+import { EntityStoreService } from '../../../services/entity-store.service';
 
 @Component({
   selector: 'app-restaurant-settings',
@@ -21,8 +22,10 @@ export class RestaurantSettingsComponent implements OnInit {
   @ViewChild('areaForm') areaForm: NgForm;
   @ViewChild('areaDetailForm') areaDetailForm: NgForm;
   @ViewChild('tableForm') tableForm: NgForm;
+  floorNames: Map<string, string> = new Map();
 
-  constructor(private mainService: MainService, private messageService: MessageService, private logService: LogService) {
+  constructor(private mainService: MainService, private messageService: MessageService,
+              private logService: LogService, private entityStoreService: EntityStoreService) {
     this.onUpdate = false;
     this.fillData();
   }
@@ -39,7 +42,7 @@ export class RestaurantSettingsComponent implements OnInit {
     this.tableForm.reset();
   }
 
-  getTablesByFloor(id) {
+  getTablesByFloor(id: string) {
     if (id) {
       this.mainService.getAllBy('tables', { floor_id: id }).then(result => {
         this.tables = result.docs;
@@ -53,21 +56,21 @@ export class RestaurantSettingsComponent implements OnInit {
     }
   }
 
-  getFloor(floor) {
+  getFloor(floor: Floor) {
     this.selectedFloor = floor._id;
     floor = Object.assign(floor, floor.conditions);
     delete floor.conditions;
     this.areaDetailForm.setValue(floor);
   }
 
-  addFloor(areaForm) {
-    let form = areaForm.value;
+  addFloor(areaForm: NgForm) {
+    const form = areaForm.value;
     if (!form.name) {
       this.messageService.sendMessage('Bölüm Adı Belirtmelisiniz');
       return false;
     }
-    let areaSpecs = new FloorSpecs(form.air, form.cigarate, form.reservation, form.music, form.events);
-    let schema = new Floor(form.name, form.description, 1, Date.now(), form.special, areaSpecs);
+    const areaSpecs = new FloorSpecs(form.air, form.cigarate, form.reservation, form.music, form.events);
+    const schema = new Floor(form.name, form.description, 1, Date.now(), form.special, areaSpecs);
     this.mainService.addData('floors', schema).then(() => {
       this.fillData();
       this.messageService.sendMessage('Bölüm Oluşturuldu!');
@@ -77,8 +80,9 @@ export class RestaurantSettingsComponent implements OnInit {
   }
 
   updateFloor(areaDetailForm: NgForm) {
-    let form = areaDetailForm.value;
-    let schema = new Floor(form.name, form.description, 1, Date.now(), form.special, new FloorSpecs(form.air, form.cigarate, form.reservation, form.music, form.events), form._id, form._rev);
+    const form = areaDetailForm.value;
+    const schema = new Floor(form.name, form.description, 1, Date.now(), form.special,
+      new FloorSpecs(form.air, form.cigarate, form.reservation, form.music, form.events), form._id, form._rev);
     this.mainService.updateData('floors', form._id, schema).then(res => {
       this.selectedFloor = undefined;
       this.fillData();
@@ -87,18 +91,21 @@ export class RestaurantSettingsComponent implements OnInit {
   }
 
   removeFloor() {
-    let isOk = confirm('Bölümü Silmek Üzeresiniz. Bölüme Dahil Olan Masalarda Silinecektir.');
+    const isOk = confirm('Bölümü Silmek Üzeresiniz. Bölüme Dahil Olan Masalarda Silinecektir.');
     if (isOk) {
       this.mainService.removeData('floors', this.selectedFloor).then(() => {
         this.mainService.getAllBy('tables', { floor_id: this.selectedFloor }).then(result => {
-          let data = result.docs
-          for (let prop in data) {
-            this.mainService.removeData('tables', data[prop]._id).then(result => {
-              this.mainService.getAllBy('reports', { connection_id: result.id }).then((res) => {
-                if (res.docs.length > 0)
-                  this.mainService.removeData('reports', res.docs[0]._id);
+          const data = result.docs;
+          for (const prop in data) {
+            if (Object.prototype.hasOwnProperty.call(data, prop)) {
+              this.mainService.removeData('tables', data[prop]._id).then(resTable => {
+                this.mainService.getAllBy('reports', { connection_id: resTable.id }).then((res) => {
+                  if (res.docs.length > 0) {
+                    this.mainService.removeData('reports', res.docs[0]._id);
+                  }
+                });
               });
-            });
+            }
           }
           this.messageService.sendMessage('Bölüm ve Masalar Silindi!')
           this.selectedFloor = undefined;
@@ -109,7 +116,7 @@ export class RestaurantSettingsComponent implements OnInit {
   }
 
   addTable(tableForm: NgForm) {
-    let form = tableForm.value;
+    const form = tableForm.value;
     if (!form.name) {
       this.messageService.sendMessage('Masa Adı Belirtmelisiniz');
       return false;
@@ -117,10 +124,12 @@ export class RestaurantSettingsComponent implements OnInit {
       this.messageService.sendMessage('Kategori Seçmelisiniz');
       return false;
     }
-    let schema = new Table(form.name, form.floor_id, form.capacity, form.description, 1, Date.now(), [], form._id, form._rev);
-    if (form._id == undefined) {
+    const schema = new Table(form.name, form.floor_id, form.capacity, form.description, 1, Date.now(), [], form._id, form._rev);
+    if (form._id === undefined) {
       this.mainService.addData('tables', schema).then((response) => {
-        this.mainService.addData('reports', new Report('Table', response.id, 0, 0, 0, [0, 0, 0, 0, 0, 0, 0], [0, 0, 0, 0, 0, 0, 0], [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], new Date().getMonth(), new Date().getFullYear(), form.name, Date.now())).then(res => {
+        this.mainService.addData('reports', new Report('Table', response.id, 0, 0, 0, [0, 0, 0, 0, 0, 0, 0],
+          [0, 0, 0, 0, 0, 0, 0], [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+          new Date().getMonth(), new Date().getFullYear(), form.name, Date.now())).then(res => {
           this.logService.createLog(logType.TABLE_CREATED, res.id, `${form.name} adlı Masa oluşturuldu.`);
         });
         this.fillData();
@@ -138,24 +147,28 @@ export class RestaurantSettingsComponent implements OnInit {
     $('#tableModal').modal('hide');
   }
 
-  updateTable(id) {
+  updateTable(id: string) {
     this.onUpdate = true;
     this.selectedTable = id;
     this.mainService.getData('tables', id).then((result: Table) => {
       delete result.customers;
-      this.tableForm.setValue(result);
+      if ((result as any).moveTable !== undefined) {
+        delete (result as any).moveTable;
+      }
+      this.tableForm.form.patchValue(result);
       $('#tableModal').modal('show');
     });
   }
 
   removeTable() {
-    let isOk = confirm('Masayı Silmek Üzeresiniz!');
+    const isOk = confirm('Masayı Silmek Üzeresiniz!');
     if (isOk) {
       this.mainService.removeData('tables', this.selectedTable).then((result) => {
         this.logService.createLog(logType.TABLE_DELETED, result._id, `${this.tableForm.value.name} adlı Masa Silindi.`);
         this.mainService.getAllBy('reports', { connection_id: result.id }).then((res) => {
-          if (res.docs.length > 0)
+          if (res.docs.length > 0) {
             this.mainService.removeData('reports', res.docs[0]._id);
+          }
         });
         this.fillData();
         this.messageService.sendMessage('Masa Silindi.');
@@ -167,7 +180,7 @@ export class RestaurantSettingsComponent implements OnInit {
   }
 
   filterTables(value: string) {
-    let regexp = new RegExp(value, 'i');
+    const regexp = new RegExp(value, 'i');
     this.mainService.getAllBy('tables', { name: { $regex: regexp } }).then(res => {
       this.tables = res.docs;
       this.tables = this.tables.sort((a, b) => a.name.localeCompare(b.name));
@@ -182,6 +195,11 @@ export class RestaurantSettingsComponent implements OnInit {
     this.mainService.getAllBy('tables', {}).then((result) => {
       this.tables = result.docs;
       this.tables = this.tables.sort((a, b) => a.name.localeCompare(b.name));
+
+      const floorIds = this.tables.map(t => t.floor_id).filter(id => id);
+      this.entityStoreService.resolveEntities('floors', floorIds).then(resolved => {
+        this.floorNames = resolved;
+      });
     });
   }
 }
